Improved memory usage Memory usage of the local noise scaling algorithm was improved by enforcing that the generated rms cube has the same data type as the original data cube rather than 'float64' by default. Otherwise the module would require an additional 2x the cube size in memory for 32-bit floating point data. The dynamic range of float32 should be sufficient to describe the noise variations across any sensible cube.

+++ sofia/sigma_cube.py
@@ -76,7 +76,7 @@ def sigma_scale(cube, scaleX=False, scaleY=False, scaleZ=True, edgeX=0, edgeY=0,
 		radiusWindowSpectral = windowSpectral // 2
 		
 		# Create empty cube (filled with NaN) to hold noise values
-		rms_cube = np.full(cube.shape, np.nan)
+		rms_cube = np.full(cube.shape, np.nan, dtype=cube.dtype)
 		
 		# Determine RMS across window centred on grid cell
 		for z in gridPointsZ:
